Finish creating migration, all done. Will add relation in model later. Laravel's migration is sucks compared to django.

+++ iduka/database/migrations/2020_10_13_152306_create_kategori_produks_table.php
@@ -16,7 +16,7 @@ class CreateKategoriProduksTable extends Migration
         Schema::create('kategori_produks', function (Blueprint $table) {
             $table->bigIncrements('kategori_produk_id');
             $table->string('nama_kategori');
-            $table->integer('umkm_id')->unsigned();
+            $table->bigInteger('umkm_id')->unsigned();
         });
 
         Schema::table('kategori_produks', function (Blueprint $table) {
